Use a full-data content hash as the texture replacement key
Cemu's texDataHash2 samples only ~296 bytes, so distinct textures could share
a key -- monster subspecies built from a common base texture collided and
received each other's replacements. Replacement lookups now use a full-data
hash of the guest mip0 surface. Texture dumps use the same key so a dumped
filename can be reused verbatim.

Breaking: pack filenames change from 8-hex to 16-hex; packs must be re-converted.

+++ src/Cafe/HW/Latte/Core/LatteTexture.cpp
@@ -1356,7 +1356,7 @@ LatteTexture::LatteTexture(Latte::E_DIM dim, MPTR physAddress, MPTR physMipAddre
 	{
 		LatteAddrLib::AddrSurfaceInfo_OUT _replSI;
 		LatteAddrLib::GX2CalculateSurfaceInfo(format, width, height, depth, dim, Latte::MakeGX2TileMode(tileMode), 0, 0, &_replSI);
-		uint32 _replHash = LatteTextureReplace::HashGuest(this->physAddress, (uint32)_replSI.surfSize, width * height, format);
+		uint64 _replHash = LatteTextureReplace::HashGuest(this->physAddress, (uint32)_replSI.surfSize, width * height, format);
 		LatteTextureReplace::ReplacementInfo _ri;
 		if (LatteTextureReplace::GetInfo(_replHash, _ri))
 		{
